+++ src/common/css-class-definition.ts
@@ -1,9 +1,7 @@
-"use strict";
-
 import * as vscode from "vscode";
 
 class CssClassDefinition {
-    public constructor(public className: string, public location?: vscode.Location) { }
+    public constructor(public className: string,public baseName:string, public location?: vscode.Location) { }
 }
 
 export default CssClassDefinition;