+++ src/parse-engines/common/simple-text-document.ts
@@ -4,6 +4,7 @@
 interface ISimpleTextDocument {
     languageId: string;
     getText(): string;
+    baseName: string;
 }
 
 export default ISimpleTextDocument;