+++ src/parse-engines/types/less-parse-engine.ts
@@ -1,19 +1,21 @@
 import * as css from "css";
+//@ts-ignore
+import less from 'less/lib/less-node';
 import CssClassDefinition from "../../common/css-class-definition";
 import CssClassExtractor from "../common/css-class-extractor";
 import IParseEngine from "../common/parse-engine";
 import ISimpleTextDocument from "../common/simple-text-document";
 
-class CssParseEngine implements IParseEngine {
-    public languageId = "css";
-    public extension = "css";
+class LessParseEngine implements IParseEngine {
+    public languageId = "less";
+    public extension = "less";
 
     public async parse(textDocument: ISimpleTextDocument): Promise<CssClassDefinition[]> {
         const code: string = textDocument.getText();
-        const codeAst: css.Stylesheet = css.parse(code);
-
+        const res:Less.RenderOutput = await less.render(code);
+        const codeAst: css.Stylesheet = css.parse(res.css);
         return CssClassExtractor.extract(codeAst,textDocument.baseName);
     }
 }
 
-export default CssParseEngine;
+export default LessParseEngine;